EFK-FE-12 refactor: move files to another folder
