+++ efk-fe/src/components/CardsContainer/CardsContainer.module.scss
@@ -1,6 +1,6 @@
 @import '../../styles/styles-config.scss';
 
-.categoryList {
+.cards {
   width: 80%;
   @include flex($justify: space-evenly);
