+++ efk-fe/src/components/CardsContainer/CardsContainer.tsx
@@ -1,0 +1,10 @@
+import { FC, ReactFragment } from 'react';
+import styles from './CardsContainer.module.scss';
+
+interface CardsContainerProps {
+  children: ReactFragment;
+}
+
+export const CardsContainer: FC<CardsContainerProps> = ({ children }) => {
+  return <section className={styles.cards}>{children}</section>;
+};
