+++ efk-fe/src/components/CardsContainer/index.ts
@@ -1,0 +1,1 @@
+export { CardsContainer } from './CardsContainer';
