+++ efk-fe/src/components/Categories/CategoryCard/CategoryCard.module.scss
@@ -1,0 +1,36 @@
+@import '../../../styles/styles-config.scss';
+
+.categoryCard {
+  @include card();
+  background: white;
+
+  &:hover {
+    background-color: rgb(255, 251, 218);
+    box-shadow: 3px 5px 6px rgb(0 0 0 / 100%);
+  }
+
+  &__img {
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    margin-bottom: 3px;
+  }
+
+  &__name {
+    color: rgb(58, 58, 58);
+    text-transform: uppercase;
+    text-align: center;
+  }
+  &__link {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    border-radius: 19px;
+    transition: all 0.4s ease-in-out;
+    z-index: 1;
+  }
+}
