--- efk-fe/src/components/CategoriesContainer/CategoryCard.module.scss
@@ -1,46 +1,0 @@
-.categoryCard {
-  position: relative;
-  margin-bottom: 40px;
-  width: 250px;
-  height: 250px;
-  transition: 0.5s ease;
-  display: flex;
-  flex-direction: column;
-  border-radius: 19px;
-  box-shadow: 19px -4px 39px -20px #cdcdcd, -14px 11px 15px -12px #cdcdcd;
-  background: white;
-  &:hover {
-    background-color: rgb(255, 251, 218);
-    box-shadow: 3px 5px 6px rgb(0 0 0 / 100%);
-  }
-}
-.categoryCard__img {
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  margin-bottom: 3px;
-}
-.categoryCard__name {
-  color: rgb(58, 58, 58);
-  text-transform: uppercase;
-  text-align: center;
-}
-.categoryCard__link {
-  position: absolute;
-  cursor: pointer;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  border-radius: 19px;
-  transition: all 0.4s ease-in-out;
-  z-index: 1;
-}
-
-@media screen and (min-width: 1500px) {
-  .categoryCard {
-    width: calc(100% / 4 - 40px);
-    height: calc(70vh / 2 - 20px);
-  }
-}
