--- efk-fe/src/components/CategoriesContainer/index.ts
@@ -1,1 +1,0 @@
-export { CategoriesContainer } from './CategoriesContainer';
