+++ efk-fe/src/components/Words/WordsContainer/WordsContainer.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/styles-config.scss';
+@import '../../../styles/styles-config.scss';
 
 .cardsContainer {
   width: 100%;
