--- efk-fe/src/components/WordsContainer/index.ts
@@ -1,1 +1,0 @@
-export { WordsContainer } from './WordsContainer';
