--- efk-fe/src/pages/category/Category.module.scss
@@ -1,6 +1,0 @@
-@import '../../styles/styles-config.scss';
-
-.category {
-  width: 80%;
-  @include flex();
-}
